Updates RecordingIsPosted

+++ CUWebinars.Business/Repository/IWebinarRepository.cs
@@ -37,6 +37,6 @@ namespace CUWebinars.Business.Repository
         int SaveChanges();
         void SynchToLegacy();
         int GetRegTypeByACS(string registrationType, int idWebinar);
-        decimal[] GetCostOfUpgrades(int idWebinar);
+        double[] GetCostOfUpgrades(int idWebinar);
     }
 }
